Rename the files to lowercase.
