+++ lib/models/people.dart
